src/bot/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
discord.py==2.3.2
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
multidict==6.0.4
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.0.1
yarl==1.9.4
psycopg[binary]
